Extract callsign parsing into pkg/callsigns and shared text normalization into internal/normalize
ParsePilotCallsign lived in pkg/parser alongside speech-recognition request parsing, forcing
pkg/radar and pkg/controller to import the full parser just for callsign handling. This separates
concerns by moving callsign parsing to its own package and extracting shared text normalization
(string cleanup, digit spacing, digit words) into internal/normalize.

Also removes the redundant strings.Cut("|") in radar/contacts.go — normalize.String() already
handles pipe-delimited input.
