+++ pkg/callsigns/callsign_test.go
@@ -1,4 +1,4 @@
-package parser
+package callsigns
 
 import (
 	"strconv"
