+++ pkg/controller/broadcast.go
@@ -3,12 +3,12 @@ package controller
 import (
 	"slices"
 
-	"github.com/dharmab/skyeye/pkg/parser"
+	"github.com/dharmab/skyeye/pkg/callsigns"
 	"github.com/dharmab/skyeye/pkg/trackfiles"
 	"github.com/rs/zerolog/log"
 )
 
-func (c *Controller) addFriendlyToBroadcast(callsigns []string, friendly *trackfiles.Trackfile) []string {
+func (c *Controller) addFriendlyToBroadcast(friendlyCallsigns []string, friendly *trackfiles.Trackfile) []string {
 	logger := log.With().Str("callsign", friendly.Contact.Name).Logger()
 	isOnFrequency := c.srsClient.IsOnFrequency(friendly.Contact.Name)
 	if isOnFrequency {
@@ -17,14 +17,14 @@ func (c *Controller) addFriendlyToBroadcast(callsigns []string, friendly *trackf
 
 	shouldBroadcast := !c.threatMonitoringRequiresSRS || isOnFrequency
 	if !shouldBroadcast {
-		return callsigns
+		return friendlyCallsigns
 	}
-	if callsign, ok := parser.ParsePilotCallsign(friendly.Contact.Name); ok {
-		if !slices.Contains(callsigns, callsign) {
-			callsigns = append(callsigns, callsign)
+	if callsign, ok := callsigns.ParsePilotCallsign(friendly.Contact.Name); ok {
+		if !slices.Contains(friendlyCallsigns, callsign) {
+			friendlyCallsigns = append(friendlyCallsigns, callsign)
 		}
 	} else {
 		logger.Debug().Msg("could not parse callsign")
 	}
-	return callsigns
+	return friendlyCallsigns
 }
